+++ js/src/front-end.js
@@ -2,7 +2,7 @@
  * @Author: Tuomas Marttila
  * @Date:   2023-02-17 10:10:38
  * @Last Modified by:   Tuomas Marttila
- * @Last Modified time: 2023-02-27 10:16:36
+ * @Last Modified time: 2023-04-12 15:36:12
  */
 /* eslint-disable max-len, no-param-reassign, no-unused-vars */
 /**
@@ -16,7 +16,7 @@ import initAnchors from './modules/anchors';
 import backToTop from './modules/top';
 import initA11ySkipLink from './modules/a11y-skip-link';
 import initA11yFocusSearchField from './modules/a11y-focus-search-field';
-import { navDesktop, navMobile } from './modules/navigation';
+import { navDesktop, navClick, navMobile } from './modules/navigation';
 
 // Define Javascript is active by changing the body class
 document.body.classList.remove('no-js');
@@ -31,7 +31,10 @@ document.addEventListener('DOMContentLoaded', () => {
   initA11yFocusSearchField();
 
   // Init navigation
+  // If you want to enable click based navigation, comment navDesktop() and uncomment navClick()
+  // Remember to enable styles in sass/navigation/navigation.scss
   navDesktop();
+  // navClick();
   navMobile();
 
   // Fit video embeds to container